fix:完善边界条件

+++ sky-server/src/main/java/com/sky/mapper/SetmealMapper.java
@@ -11,5 +11,5 @@ public interface SetmealMapper {
      * @return
      */
     @Select("select count(*) from setmeal where category_id = #{categoryId}")
-    Integer countByCategoryId(Integer categoryId);
+    Integer countByCategoryId(Long categoryId);
 }
